requirements.txt
astroid==2.4.2
autopep8==1.5.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.5.3
python-dotenv==0.14.0
rope==0.17.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
